feat:  implement UvcServer to run uvc in standalone mode and trim UvcStreamer and implement helper_fill_mjpg_image
fixme: python `app` runtime was not waiting for destructors to finish before terminating the process, but just sys.exit(0) in signal_handle in this file `/usr/lib/python3.11/site-packages/maix/__signal_handle__.py`
