+++ components/vision/port/maixcam/uvc-gadget-server.c
