+++ examples/uvc_demo/main/src/main.cpp
@@ -3,10 +3,6 @@
 
 using namespace std;
 
-#include <termios.h>
-#include <unistd.h>
-#include <fcntl.h>
-
 #include "main.h"
 #include "maix_util.hpp"
 #include "maix_image.hpp"
@@ -18,54 +14,42 @@ using namespace std;
 
 using namespace maix;
 
-#include "uvcServer.hpp"
+using namespace maix::uvc;
 
 static std::unique_ptr<camera::Camera> pCam = nullptr;
-// extern "C" 
 int my_uvc_video_fill_mjpg_buffer(void *buf, uint32_t *size) {
     static uint64_t frame_count = 0;
     static uint64_t last_ms;
     if (!pCam) return 1;
 
-uint64_t curr_ms_thisframe = time::ticks_ms();
     image::Image *img = pCam->read();
     if (!img) return 1;
-uint64_t curr_ms_takeframe = time::ticks_ms();
+    uint64_t curr_ms_takeframe = time::ticks_ms();
+
     std::ostringstream oss;
     oss << "frame: " << frame_count;
     img->draw_string(4, 32, oss.str(), image::Color::from_rgb(255, 0, 0), 1.3);
 
-    image::Image *img_mjpg = img;
-    if (img->format() != image::Format::FMT_JPEG) {
-        img_mjpg = img->to_jpeg(95);
-        delete img;
-    }
-    if (!img_mjpg) return 1;
-uint64_t curr_ms_mjpgframe = time::ticks_ms();
+    uvc::helper_fill_mjpg_image(buf, size, img);
+    uint64_t curr_ms_mjpgframe = time::ticks_ms();
 
-    memcpy(buf, img_mjpg->data(), img_mjpg->data_size());
-    *size = img_mjpg->data_size();
+    delete img;
 #if 0
     char result_bin_name[32];
     sprintf((char *)result_bin_name, "/root/res/res_%llu.jpg", frame_count);
     ofstream((const char*)result_bin_name, std::ios::binary)
-        .write(reinterpret_cast<char*>(img_mjpg->data()), img_mjpg->data_size());
+        .write(reinterpret_cast<char*>(buf), *size);
 #endif
-    delete img_mjpg;
 
     uint64_t curr_ms = time::ticks_ms();
     static uint32_t flip = 1;
     if(flip^=1) log::info("[%llu]loop use %lld ms, %.2ffps, %d bytes\r\n"
-                        "\tstart: %lld ms\r\n"
                         "\ttake: %lld ms\r\n"
                         "\tto_mjpg: %lld ms\r\n", frame_count++, curr_ms - last_ms, 1000.f/(curr_ms - last_ms), *size,
-                        curr_ms_thisframe - last_ms, curr_ms_takeframe - curr_ms_thisframe, curr_ms_mjpgframe - curr_ms_takeframe);
+                        curr_ms_takeframe - last_ms, curr_ms_mjpgframe - curr_ms_takeframe);
     last_ms = curr_ms;
     return 0;
 }
-
-// extern "C" int uvc_main_exist;
-// extern "C" int uvc_main(int argc, char *argv[]);
 
 int _main(int argc, char* argv[])
 {
@@ -92,22 +76,12 @@ int _main(int argc, char* argv[])
     camera::Camera cam = camera::Camera(cam_w, cam_h, cam_fmt, "", cam_fps, cam_buffer_num);
     if(!pCam) pCam = make_unique<camera::Camera>(std::move(cam));
 
-    // std::thread uvc_daemon_thrd([&argv] () {
-    //     log::info("Switch to uvc_main...");
-    //     char* uvc_main_args[] = {argv[0], "-u", "/dev/video0", "-d", "-i", "/bin/cat_224.jpg"};
-    //     uvc_main(sizeof(uvc_main_args)/sizeof(uvc_main_args[0]), uvc_main_args);
-    // });
-
-    UvcServer *pUvc = UvcServer::getInstance();
-    pUvc->set_cb(my_uvc_video_fill_mjpg_buffer);
+    std::unique_ptr<UvcServer> pUvc = std::make_unique<UvcServer>(my_uvc_video_fill_mjpg_buffer);
     pUvc->run();
     while (!app::need_exit()) {
         std::this_thread::sleep_for(std::chrono::seconds(1));
     }
     pUvc->stop();
-    // uvc_main_exist = 1;
-    
-    // uvc_daemon_thrd.join();
 
     return 0;
 }
